+++ src/main/java/com/dduongdev/hotel/exception/RoomAlreadyBookedException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 
 public class RoomAlreadyBookedException extends BaseRoomException {
 
-    public RoomAlreadyBookedException(Integer roomId, String checkIn, String checkOut) {
+    public RoomAlreadyBookedException(Long roomId, String checkIn, String checkOut) {
         super(HttpStatus.BAD_REQUEST, "Room with id " + roomId + " is already booked for the period from " + checkIn + " to " + checkOut);
     }
 }
