+++ src/main/java/com/dduongdev/hotel/payload/request/UpdateRoomRequest.java
@@ -17,5 +17,5 @@ public class UpdateRoomRequest {
     private boolean hidden;
 
     @NotNull(message = "Room type ID is required")
-    private int roomTypeId;
+    private Long roomTypeId;
 }
